requirements.txt
boto==2.49.0
boto3==1.9.66
botocore==1.12.66
bz2file==0.98
certifi==2018.11.29
chardet==3.0.4
docutils==0.14
emoji==0.5.1
gensim==3.6.0
idna==2.8
Janome==0.3.7
jmespath==0.9.3
mecab-python3==0.996.1
mojimoji==0.0.8
mysql-connector-python==8.0.13
numpy==1.15.4
oauthlib==2.1.0
pprint==0.1
protobuf==3.6.1
python-dateutil==2.7.5
python-dotenv==0.10.1
requests==2.21.0
requests-oauthlib==1.0.0
s3transfer==0.1.13
scipy==1.1.0
six==1.12.0
smart-open==1.7.1
urllib3==1.24.1
